pyproject.toml
[project]
name = "brale-cli"
version = "0.1.0"
description = "Brale API command-line interface"
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "pyyaml>=6.0.0",
    "tabulate>=0.9.0",
]
requires-python = ">=3.8"

[project.scripts]
brale = "brale.cli:main"

[tool.uv]
package = true